requirements.txt
torch>=1.9.0
numpy>=1.21.0
plotly>=5.3.0
scikit-image>=0.18.0
pyyaml>=5.4.0
pytest>=6.2.0
mypy>=0.910
matplotlib>=3.4.0
pandas>=1.3.0
trimesh>=3.9.0
scipy>=1.7.0
